RepositorysCompletos

+++ gestionart_frontend_ruben_y_jessica/lib/data/repositories/AnuncioRepository.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:gestionart_frontend_ruben_y_jessica/data/models/Anuncio.dart';
 import 'package:gestionart_frontend_ruben_y_jessica/services/ApiService.dart';
 
@@ -7,7 +9,7 @@ class Anunciorepository {
   Anunciorepository(ApiService? apiService) : _apiService = apiService ?? ApiService();
 
 
-  Future<bool> crearAnuncio(int idVendedor, String titulo, String categoria, double precio, String imagen) async {
+  Future<bool> crearAnuncio(Long idVendedor, String titulo, String categoria, double precio, String imagen) async {
     try {
       final response = await _apiService.dio.post("http://localhost:8080/anuncios", data: {
         "idVendedor": idVendedor,
@@ -27,7 +29,7 @@ class Anunciorepository {
   
   }
   
-  Future<Anuncio> getAnuncioPorVendedor(int idVendedor) async {
+  Future<Anuncio> getAnuncioPorVendedor(Long idVendedor) async {
     try {
       final response = await _apiService.dio.get("http://localhost:8080/anuncios/vendedor/$idVendedor");
       if (response.statusCode == 200) {
@@ -57,7 +59,7 @@ class Anunciorepository {
     }
   }
 
-  Future<Anuncio> getAnuncioPorId(int id) async {
+  Future<Anuncio> getAnuncioPorId(Long id) async {
     try {
       final response = await _apiService.dio.get("http://localhost:8080/anuncios/$id");
       if (response.statusCode == 200) {
@@ -83,7 +85,7 @@ class Anunciorepository {
     }
   }
 
-  Future<void> eliminarAnuncio(int id) async {
+  Future<void> eliminarAnuncio(Long id) async {
     try {
       final response = await _apiService.dio.delete("http://localhost:8080/anuncios/$id");
       if (response.statusCode != 200) {
