+++ gestionart_frontend_ruben_y_jessica/lib/data/repositories/VendedorRepository.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:gestionart_frontend_ruben_y_jessica/data/models/Vendedor.dart';
 import 'package:gestionart_frontend_ruben_y_jessica/services/ApiService.dart';
 
@@ -6,7 +8,7 @@ class Vendedorrepository {
   final ApiService apiService;
   Vendedorrepository(ApiService? apiService) : apiService = apiService ?? ApiService();
 
-  Future<void> actualizarVendedor(int id, String correoElectronico, String nombre, String descripcionPerfil, String imagen, String contrasena) async {
+  Future<void> actualizarVendedor(Long id, String correoElectronico, String nombre, String descripcionPerfil, String imagen, String contrasena) async {
     try {
       final response = await apiService.dio.put("http://localhost:8080/vendedores/$id", data: {
         "correoElectronico": correoElectronico,
@@ -39,7 +41,7 @@ class Vendedorrepository {
       throw Exception("Error al obtener los vendedores: $e");
     }
   }
-  Future<Vendedor> getVendedorPorId(int id) async {
+  Future<Vendedor> getVendedorPorId(Long id) async {
     try {
       final response = await apiService.dio.get("http://localhost:8080/vendedores/$id");
       if (response.statusCode == 200) {
@@ -78,7 +80,7 @@ class Vendedorrepository {
     }
   }
 
-  Future<void> eliminarVendedor(int id) async {
+  Future<void> eliminarVendedor(Long id) async {
     try {
       final response = await apiService.dio.delete("http://localhost:8080/vendedores/$id");
       if (response.statusCode != 200) {
